# src/main/resources/openapi.yaml
openapi: 3.0.0
servers:
  - description: Local Development Server
    url: http://localhost:8911
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/philippsuniversity/CarSharingAPI/1.0.0
info:
  description: |
    B2B Carpooling Application API for managing user accounts, vehicles, ride offers, 
    and ride requests for a corporate carpooling solution.
  version: "1.0.0"
  title: B2B Carpooling API
  contact:
    email: your.email@university.edu
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
  /api/users/signup:
    post:
      tags:
        - users
      summary: Register a new user
      operationId: registerUser
      description: Creates a new user account in the system
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSignupRequest'
            examples:
              userSignup:
                value:
                  requestContent:
                    name: "John Doe"
                    EmailId: "john.doe@company.com"
                    phone: "1234567890"
                    password: "securePassword123"
        required: true
      responses:
        '200':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupResponse'
              examples:
                success:
                  value:
                    status: true
                    message: "User registered successfully"
                    responseContent:
                      userId: "user123"
                      name: "John Doe"
                      email: "john.doe@company.com"
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/login:
    post:
      tags:
        - users
      summary: User login
      operationId: loginUser
      description: Authenticates a user and returns login details
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              login:
                value:
                  requestContent:
                    email: "john.doe@company.com"
                    password: "securePassword123"
        required: true
      responses:
        '200':
          description: User successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                success:
                  value:
                    status: true
                    message: "Login successful"
                    responseContent:
                      userId: "user123"
                      name: "John Doe"
                      email: "john.doe@company.com"
                      token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    RequestDTO:
      type: object
      properties:
        userId:
          type: string
          example: "user123"
        requestContent:
          type: object
          description: "Content varies based on endpoint"
    ResponseDTO:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
        responseContent:
          type: object
          description: "Content varies based on endpoint"
    ErrorResponse:
      type: object
      properties:
        status:
          type: boolean
          example: false
        message:
          type: string
          example: "Error occurred"
        errorCode:
          type: string
          example: "ERR_INVALID_INPUT"

    # User-related schemas
    UserInfoDTO:
      type: object
      required:
        - name
        - emailid
        - password
      properties:
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@company.com"
        phone:
          type: string
          example: "1234567890"
        password:
          type: string
          format: password
          example: "securePassword123"
    UserSignupRequest:
      allOf:
        - $ref: '#/components/schemas/RequestDTO'
        - type: object
          properties:
            requestContent:
              $ref: '#/components/schemas/UserInfoDTO'
    SignupResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseDTO'
        - type: object
          properties:
            responseContent:
              type: object
              properties:
                userId:
                  type: string
                  example: "user123"
                name:
                  type: string
                  example: "John Doe"
                email:
                  type: string
                  example: "john.doe@company.com"
    LoginRequestDTO:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "john.doe@company.com"
        password:
          type: string
          format: password
          example: "securePassword123"
    LoginRequest:
      allOf:
        - $ref: '#/components/schemas/RequestDTO'
        - type: object
          properties:
            requestContent:
              $ref: '#/components/schemas/LoginRequestDTO'
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseDTO'
        - type: object
          properties:
            responseContent:
              type: object
              properties:
                userId:
                  type: string
                  example: "user123"
                name:
                  type: string
                  example: "John Doe"
                email:
                  type: string
                  example: "john.doe@company.com"
                token:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
...